[fix] n double

+++ 0x1E-search_algorithms/100-jump.c
@@ -16,7 +16,7 @@ int jump_search(int *array, size_t size, int value)
 	if (array == NULL || size == 0)
 		return (-1);
 
-	n = sqrt((double)size);
+	n = sqrt(size);
 	for (i = jp = 0; jp < (int)size && array[jp] < value;)
 	{
 		printf("Value checked array[%d] = [%d]\n", i, array[jp]);
